レビュー指摘など修正
・FlurryLibraryのversionを8.2.0に戻した
・logEventのparamの型をint→longに変更
・Flurry→FlurryAnalyticsToolStrategy. をインタフェースから実装するように修正
・車載器連携開始と連携終了で投げるイベントのメソッドをまとめた
・視聴ソース情報収集計測は音声認識などでの一時的ソース変更で停止し、元ソースに戻ると再開するように修正。
・ソース切り替え操作トリガーセットの1秒後クリアはしない
・ソース切り替え操作トリガー送信で非対応ソースを無視する場合もトリガーをクリア。
・10秒後ソース固定でトリガー送信するのは車載器トグル、カスタムキートグルのみソース選択トリガーでは即送信するように修正。
・sThirdAppStartUpSendFlgをセットリストに修正。
・アプリ状態変更とソース切り替えとAudioMode切り替えのイベントはAnalyticsクラス内で受け取るように修正。
・フォアグラウンド復帰時の画面滞留時間開始はPause前にsCurrentScreenに保存した表示画面がAV/HOME画面であれば再開。

・3rdApp起動で一時的ソース変更トリガーをセットしていたのを修正。

+++ app/src/main/java/jp/pioneer/carsync/application/content/AnalyticsEventSubmitter.java
@@ -2,7 +2,7 @@ package jp.pioneer.carsync.application.content;
 
 public interface AnalyticsEventSubmitter {
     AnalyticsEventSubmitter with(Analytics.AnalyticsParam param, String value);
-    AnalyticsEventSubmitter with(Analytics.AnalyticsParam param, int value);
+    AnalyticsEventSubmitter with(Analytics.AnalyticsParam param, long value);
     AnalyticsEventSubmitter with(Analytics.AnalyticsParam param, double value);
     void submit();
 }
